Only clear the android_url table when it exists.
BUG=
TEST=Passed the existing tests.


Review URL: http://codereview.chromium.org/10209031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@135476 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6315d166de217368aac830d79d0d26a6e408adef

+++ chrome/browser/history/android/android_urls_database.cc
@@ -141,7 +141,12 @@ bool AndroidURLsDatabase::UpdateAndroidURLRow(AndroidURLID id,
 }
 
 bool AndroidURLsDatabase::ClearAndroidURLRows() {
-  return GetDB().Execute("DELETE FROM android_urls");
+  // The android_urls table might not exist if the Android content provider is
+  // never used, especially in the unit tests. See http://b/6385692.
+  if (GetDB().DoesTableExist("android_urls"))
+    return GetDB().Execute("DELETE FROM android_urls");
+
+  return true;
 }
 
 bool AndroidURLsDatabase::MigrateToVersion22() {
